🧹 refactor: Star background
